requirements.txt
langchain==0.0.208 
openai==0.27.8
python-dotenv 
tiktoken
deeplake
pypdf

layers/